refactor: ArticleList 컴포넌트 파일 분리 (베스트와 일반 게시글 리스트 통합)
