+++ components/boards/ArticleSection.module.scss
@@ -38,28 +38,6 @@
     display: flex;
     gap: 16px;
   }
-
-  .list {
-    margin-top: 24px;
-
-    display: flex;
-    flex-direction: column;
-    gap: 24px;
-
-    .divider {
-      margin-top: 24px;
-      border: 1px solid $cool-gray200;
-    }
-  }
-}
-
-.spinner_container {
-  width: 100%;
-  height: 217px;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
 }
 
 @media (max-width: 767px) {
