+++ components/boards/BestArticleSection.module.scss
@@ -10,44 +10,12 @@
     line-height: 1.2;
     color: $cool-gray900;
   }
-
-  .list {
-    margin-top: 24px;
-
-    display: flex;
-    gap: 24px;
-
-    li {
-      flex-grow: 1;
-    }
-  }
-}
-
-.spinner_container {
-  width: 100%;
-  height: 217px;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-@media (max-width: 1199px) {
-  .section {
-    .list {
-      gap: 16px;
-    }
-  }
 }
 
 @media (max-width: 767px) {
   .section {
     .title {
       font-size: 18px;
-    }
-
-    .list {
-      margin-top: 16px;
     }
   }
 }
